// apps/api/src/modules/deals/deal-status.enum.ts
export enum DealStatus {
  DRAFT = 'DRAFT',
  WAITING_BUYER = 'WAITING_BUYER',
  WAITING_PAYMENT = 'WAITING_PAYMENT',
  FUNDS_SECURED = 'FUNDS_SECURED',
  WAITING_SHIPMENT = 'WAITING_SHIPMENT',
  SHIPPED = 'SHIPPED',
  DELIVERED = 'DELIVERED',
  INSPECTION = 'INSPECTION',
  COMPLETED = 'COMPLETED',
  PROBLEM_REPORTED = 'PROBLEM_REPORTED',
  CANCELLED = 'CANCELLED',
  EXPIRED = 'EXPIRED',
  WAITING_LEGAL_RESOLUTION = 'WAITING_LEGAL_RESOLUTION'
}